refactor!: Combine v2 request & public; Simplified code
feat: Add `is_v1` argument to `Route`.
refactor!: Make http attr as public
