+++ anilibria/api/http/__init__.py
@@ -1,4 +1,4 @@
 from .client import *
-from .v2 import *
+from .public import *
 from .public import *
 from .request import *
