+++ anilibria/api/http/client.py
@@ -1,13 +1,9 @@
-from .request import Request
-from .v2 import V2Request
 from .public import PublicRequest
 
 
 __all__ = ("HTTPClient", )
 
 
-class HTTPClient:
-    def __init__(self, proxy: str = None) -> None:
-        self.request = Request(proxy)
-        self.v2 = V2Request(self.request)
-        self.public = PublicRequest(self.request)
+class HTTPClient(PublicRequest):
+    def __init__(self, proxy: str | None = None) -> None:
+        PublicRequest.__init__(self, proxy)
